Remove redundant beans. Make method parameters final. Change history max read record count type from long to int.

+++ toc-persistence/src/main/java/com/siemens/cto/aem/persistence/dao/HistoryDao.java
@@ -14,6 +14,6 @@ public interface HistoryDao {
 
     void createHistory(String name, JpaGroup group, String event, String user);
 
-    List<JpaHistory> findHistory(String groupName, long numOfRecs);
+    List<JpaHistory> findHistory(String groupName, int numOfRecs);
 
 }
